progress: 6.34% of code matched (602/8216 functions, 541/935 objects)

+++ src/game/game_fn_800169FC.c
@@ -24,7 +24,7 @@ extern SecondEntry lbl_802417D0[];
 extern void* lbl_8024E388[];
 extern s32 fn_8016A598(void*);
 extern void fn_80163BB4(void*, const char*, ...);
-extern void fn_801E7974(void*, s16);
+extern void fn_801E7974(void*, int);
 
 s32 fn_800169FC(void* script)
 {
